Updated commands naming.
